requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
elastic-transport==9.1.0
elasticsearch==9.1.0
fastapi==0.115.12
frozenlist==1.7.0
h11==0.16.0
idna==3.10
multidict==6.6.4
neo4j==5.28.1
propcache==0.3.2
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.34.2
yarl==1.20.1
